+++ Sources/Server/Client.h
@@ -61,7 +61,7 @@ public:
 	bool m_bIsClientConnected;
 	uint32_t m_dwLastMsgId;
 	uint32_t m_dwLastMsgTime;
-	uint32_t m_dwLastMsgSize;
+	size_t m_dwLastMsgSize;
 	uint32_t m_dwLastFullObjectId;
 	uint32_t m_dwLastFullObjectTime;
